% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/teddy.R
\name{teddy}
\alias{teddy}
\title{Returns and optionally prints a teddy bear ASCII art.}
\usage{
teddy(do_print = TRUE)
}
\arguments{
\item{do_print}{A logical. Whether to print the teddy bear (Default = TRUE).}
}
\value{
str - A teddy bear ASCII art.
}
\description{
Returns and optionally prints a teddy bear ASCII art.
}
\examples{
    teddy()
}
